+++ CHANGELOG.md
@@ -79,6 +79,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     commands in `fatbuildrctl` manpage.
 
 ### Changed
+- cli: Watch task output by default when submitting tasks. The `-w, --watch`
+  option is replaced by the opposite `--batch` option to submit tasks in
+  background (#123).
 - docs:
   - Update table of available remote features with new registry content
     listing and artifact deletion possibilities.